requirements.txt
Jinja2
watchdog
python-jenkins